fix(metrics): 修复模型计数逻辑并过滤系统模型
- 将模型计数数据类型从 int 改为 long
- 在 MetricsService 中注入 ProjectService 并使用其解析数据库名称
- 修改模型计数逻辑以过滤掉系统模型
- 从 ModelService 中移除废弃的 count 方法
- 更新 ModelFmRepository 中的计数实现以过滤系统模型

+++ flexmodel-server/src/main/java/dev/flexmodel/metrics/dto/FmMetricsResponse.java
@@ -19,7 +19,7 @@ public class FmMetricsResponse {
   /**
    * 模型数量
    */
-  private int modelCount;
+  private long modelCount;
   /**
    * 分支数量
    */
